make lux hub add some space between chunks of data, minor bugfixes all around, everything seems to work

+++ led_strip_firmware/lib/lux/src/lux.c
@@ -59,7 +59,7 @@ static uint8_t cobs_remaining;
 static uint8_t cobs_add_zero;
 
 // Buffer and pointers for the COBS encoder to hold the current block
-static uint8_t cobs_encoder_fill_ptr;
+static uint16_t cobs_encoder_fill_ptr;
 static uint8_t cobs_encoder_send_ptr;
 static uint8_t cobs_buffer[256];
